style: add pre-commit hooks and editorconfig
Without wanting to overload the pre-commit hooks, this adds a few checks
on the commits to ensure validity of files.

More time-consuming checks for linting and formatting are done pre-push
to avoid impacting the developer experience during local development.
These are covered by `prettier` (for markdown, yaml, json, ...), `black`
for Python and `ruff` for Python linting.

To help standardise formatting in the future, an `.editorconfig` file
has also been added to the repository.

By virtue of using `pre-commit`, only modified files are checked which
ensures older files are incrementally updated.

Signed-off-by: JP-Ellis <josh@jpellis.me>