chore(example): migrate consumer example
Migrate the consumer example to the new example structure. This should
help reduce the redundancy between the examples and make it easier for
users to test them out and see how they work.

Ultimately, the idea will be that the consumer tests run first, and then
the provider tests run against the generated pact file.

Signed-off-by: JP-Ellis <josh@jpellis.me>