chore(build): update packaging to build ffi
As we will transition to using the Rust Pact library, we need to
download it as part of the build process. This commit adds a step to the
build process to download the library and extract it to the correct
location and then build the Python bindings.

As the Rust library is available for more platforms than the Ruby
executables, failing to find the Ruby executables is no longer a fatal
error and will instead be raised as a warning during the build process.

So small adjustments to the build script were made to accommodate this
change.

Signed-off-by: JP-Ellis <josh@jpellis.me>